Support deep copy format that I'm seeing in XCode 16. (#14)
This makes changes to support the deep copy format that I'm seeing after
I installed instrumentsToPprof for the first time today and tried to use
it with Instruments from XCode 16.  I'm guessing that maybe the format
changed recently, but I'm not sure since I've never used this before.

This makes the necessary changes for things to work for me, but doesn't
really do anything to enforce the format being consistently one format
or the other, which might be worth doing.